Add deletion of harvester after confirmation + small fixes

+++ client/src/app/indices/indices-list/indices-list.component.ts
@@ -2,6 +2,9 @@ import {Component, OnInit} from '@angular/core';
 import {IndicesService} from '../indices.service';
 import {Observable} from 'rxjs';
 import {Index} from '@shared/index.model';
+import {tap} from 'rxjs/operators';
+import {ConfirmDialogComponent} from '../../shared/confirm-dialog/confirm-dialog.component';
+import {MatDialog} from '@angular/material';
 
 @Component({
   selector: 'app-indices-list',
@@ -11,7 +14,7 @@ import {Index} from '@shared/index.model';
 export class IndicesListComponent implements OnInit {
   indices: Observable<Index[]>;
 
-  constructor(private indicesService: IndicesService) {
+  constructor(private dialog: MatDialog, private indicesService: IndicesService) {
   }
 
   ngOnInit() {
@@ -19,12 +22,19 @@ export class IndicesListComponent implements OnInit {
   }
 
   deleteIndex(name: string) {
-    this.indicesService.deleteIndex(name).subscribe(() => {
-      this.updateIndices();
+    this.dialog.open(ConfirmDialogComponent, {data: 'Wollen Sie diesen Index wirklich löschen?'}).afterClosed().subscribe(result => {
+      if (result) {
+        this.indicesService.deleteIndex(name).subscribe(() => {
+          this.updateIndices();
+        });
+      }
     });
   }
 
   private updateIndices() {
-    this.indices = this.indicesService.get();
+    this.indices = this.indicesService.get()
+      .pipe(
+        tap(indices => indices.sort((a, b) => a.name.localeCompare(b.name)))
+      );
   }
 }
